test(sdpa): the Rubin exact-LSE tests build their reference in fp64; assert_close_fp8_grad proves a single fp8 flip by its shape
PR #1059's own sm107 pipeline (67882897, job 439855192) left four cases red on the
212-SM lane, none of them a kernel defect:

* `test_fp8_stats_is_the_exact_softmax_lse[*causal*]` x3 -- max |dLSE| 1.3e-4 against the
  1e-4 bound, rms 1e-5, IDENTICAL across the f32 and f16x2 exponent arms.  The arm-
  independence is the tell: the test's own `q.float() @ k.float().T` reference runs in TF32
  in the DLFW container (`TORCH_ALLOW_TF32_CUBLAS_OVERRIDE=1`; torch also defaults
  `allow_tf32=True` in this venv), ~3e-4 relative per logit.  A 1024-column log-sum-exp
  averages that to ~2e-5; a causal row with ONE valid column keeps it whole (measured on
  A100: TF32 1.9e-4 on row 0, true fp32 1.5e-7).  Both exact-LSE tests (fp8 and mxfp8)
  now build the logits from fp64 operands, as `fp8_ref.compute_ref_forward(dtype=float64)`
  already does for the harness.  The kernels are unchanged and were exact.

* `test_mhas_v2::test_sdpa_fp8_bwd_L0[test310]` (d192x128, SWA, FROST fwd + native bwd) --
  dK row 1464 moved UNIFORMLY by +-0.5 (147 of 192 elements outside tolerance), dQ row 575
  by <= 0.25: one flipped fp8 dS code at (575, 1464) seen through Q and through K.  Row 575
  is a negative-score q row, amplitude m = 8 at d192 / attn_scale 1/8, so one dS step of
  2^-4 x dP_descale moves the whole dK row by 0.5 -- above the harness's fixed `4 * atol`
  row cap (0.32), which was calibrated on amplitude-4 rows (the GB300 test310 flip, 0.25).
  Absent from the previous nightly and passing on the 208-SM dev node (Philox layout
  follows the SM count): #1059's exact Stats re-rolled every midpoint.

  Rather than raise the cap (which is what rejects a real defect -- #981 measured 0.42-2.0),
  `assert_close_fp8_grad` gains an opt-in structure check: with `operand=` (Q for dK, K for
  dQ, dO for dV, V for O) and `flip_unit=` (dP_descale / s_descale) every over-cap bad row
  must be ONE flipped intermediate -- a rank-1 update along a SINGLE operand row whose
  fitted step is a power of two times the descale, after which the row meets the ordinary
  tolerance.  A tile, a band, a garbage row, a non-power-of-two step or a row fed by two
  flips all fail it (verified on the dumped test310 dev-node tensors and pinned by the new
  `sdpa/test_fp8_flip_budget.py`, L0).  `exec_sdpa_fp8` passes the operands at its four
  compare sites; the element and row budgets are untouched.

test/AGENTS.md records both traps (fp64 references below ~1e-3; check a flip's shape
before touching a cap).  Validated on the sm107 dev node (208 SMs, cc 10.7): test310,
fp8_bwd_ragged test31, fp8_fwd test69 / test136, both exact-LSE tests and the self-test --
19 passed, 2 skipped; self-test 5/5 on A100 too.

+++ test/python/sdpa/frost/test_sdpa_fp8_sm107.py
@@ -736,12 +736,16 @@ def test_fp8_stats_is_the_exact_softmax_lse(d_qk, d_v, causal, half_softmax):
         torch.cuda.synchronize()
         outs[with_stats] = out.clone()
     assert torch.equal(outs[True], outs[False]), "O must not depend on whether Stats is requested (Sigma normalizes O in both specializations)"
-    # The exact LSE of the problem the kernel saw: dequantized fp8 Q/K, fp32 logits.
+    # The exact LSE of the problem the kernel saw: dequantized fp8 Q/K, fp64 logits.  fp64, not fp32: the
+    # DLFW containers run fp32 matmul in TF32 (TORCH_ALLOW_TF32_CUBLAS_OVERRIDE=1, and torch's own default
+    # on Blackwell+), a ~3e-4 relative error per logit that a 1024-column log-sum-exp averages down to
+    # ~2e-5 but a causal row with one valid column keeps whole -- 1.3e-4 on the sm107 CI lane, read as a
+    # kernel failure (2026-09-15).  cf. fp8_ref.compute_ref_forward's ``dtype`` note.
     rep = hq // hkv
-    logits = (q8.float() * dq) @ (k8.float() * dk).repeat_interleave(rep, 1).transpose(-1, -2) * d_qk**-0.5
+    logits = (q8.double() * dq.double()) @ (k8.double() * dk.double()).repeat_interleave(rep, 1).transpose(-1, -2) * d_qk**-0.5
     if causal:
         logits = logits.masked_fill(~torch.tril(torch.ones(s, s, dtype=torch.bool, device=dev)), float("-inf"))
-    lse_ref = torch.logsumexp(logits.double(), dim=-1).float()
+    lse_ref = torch.logsumexp(logits, dim=-1).float()
     assert torch.isfinite(lse).all(), "unwritten LSE rows"
     err = (lse - lse_ref).abs()
     assert (
